The person you talk to can't act for the rest of the turn (#91)
In most situations, this does not change the previous behavior since we stopped advancing the turn when the conversation start and the Hero has less initiative than pretty much everyone. This means that people will have acted by the time you initiation the conversation, so you do not steal their action. Still, it's good to have, just in case we start having ways to boost initiative, like a speed potion or something.